Finish the project

+++ app/src/main/java/com/haki/storyapp/apiService/ApiService.kt
@@ -44,8 +44,8 @@ interface ApiService {
     suspend fun uploadStr(
         @Part file: MultipartBody.Part,
         @Part("description") description: RequestBody,
-        @Part("lat") lat: Float? = null,
-        @Part("lon") lon: Float? = null
+        @Part("lat") lat: Double? = null,
+        @Part("lon") lon: Double? = null
     ): UploadResponse
 
     @GET("stories/{id}")
